Refactoring: moved classifier architectures into deeplearning. Moved base classifier to base/estimators/_classifier.py. Tests pass.
