+++ sktime_dl/deeplearning/__init__.py
@@ -1,9 +1,7 @@
-from ._cnn import CNNClassifier
 from ._encoder import EncoderClassifier
 from ._fcn import FCNClassifier
 from ._mcdcnn import MCDCNNClassifier
 from ._mcnn import MCNNClassifier
-from ._mlp import MLPClassifier
 from ._resnet import ResNetClassifier
 from ._tlenet import TLENETClassifier
 from ._twiesn import TWIESNClassifier
