+++ sktime_dl/deeplearning/_encoder.py
@@ -4,7 +4,7 @@ import keras
 import keras_contrib
 import numpy as np
 
-from sktime_dl.classifiers.deeplearning._base import BaseDeepClassifier
+from sktime_dl.base.estimators._classifier import BaseDeepClassifier
 
 
 class EncoderClassifier(BaseDeepClassifier):
