+++ sktime_dl/deeplearning/_mcnn.py
@@ -10,7 +10,7 @@ import gc
 
 from sklearn.model_selection import train_test_split
 
-from sktime_dl.classifiers.deeplearning._base import BaseDeepClassifier
+from sktime_dl.base.estimators._classifier import BaseDeepClassifier
 
 
 class MCNNClassifier(BaseDeepClassifier):
