+++ sktime_dl/deeplearning/_resnet.py
@@ -3,7 +3,7 @@ __author__ = "James Large"
 import keras
 import numpy as np
 
-from sktime_dl.classifiers.deeplearning._base import BaseDeepClassifier
+from sktime_dl.base.estimators._classifier import BaseDeepClassifier
 
 
 class ResNetClassifier(BaseDeepClassifier):
