+++ sktime_dl/deeplearning/_tlenet.py
@@ -3,7 +3,7 @@ __author__ = "Aaron Bostrom, James Large"
 import keras
 import numpy as np
 
-from sktime_dl.classifiers.deeplearning._base import BaseDeepClassifier
+from sktime_dl.base.estimators._classifier import BaseDeepClassifier
 
 
 class TLENETClassifier(BaseDeepClassifier):
