+++ sktime_dl/deeplearning/cnn/__init__.py
@@ -1,0 +1,1 @@
+from ._cnn import CNNClassifier
