+++ sktime_dl/deeplearning/mlp/__init__.py
@@ -1,0 +1,1 @@
+from ._mlp import MLPClassifier
